Feature: Create Role

+++ src/shared/infra/typeorm/migrations/1651705270908-CreateTableCargos.ts
@@ -8,7 +8,7 @@ export class CreateTableCargos1651705270908 implements MigrationInterface {
                 columns: [
                     {
                         name: "id",
-                        type: "numeric",
+                        type: "serial",
                         isPrimary: true,
                     },
                     {
